Add StatusIndicator as Atlantis top level component from StatusLabel
