+++ docs/components/StatusIndicator/StatusIndicator.stories.mdx
@@ -1,0 +1,48 @@
+import { Canvas, Meta } from "@storybook/addon-docs";
+import { StatusIndicator } from "@jobber/components/StatusIndicator";
+import { Content } from "@jobber/components/Content";
+
+<Meta title="Components/Status and Feedback/StatusIndicator/Docs" />
+
+# Status Indicator
+
+The StatusIndicator component is a purely visual component that indicates a
+status (e.g., "Active", "Overdue").
+
+## Usage guidelines
+
+StatusIndicator offers only 5 possible status representations:
+
+### Success
+
+Convey that an item is in a successful state, such as approved or paid.
+
+### Critical
+
+Alert the user to a critically important issue such as a late appointment or an
+overdue payment.
+
+### Warning
+
+Warn the user of a potential forthcoming issue like an upcoming deadline or an
+item awaiting response.
+
+### Informative
+
+Inform the user about something that may not require action, such as a quote
+that has already been converted into a job.
+
+### Inactive
+
+Signify that an item has been archived, closed, or otherwise removed from an
+active workflow.
+
+## Status Indicators
+
+| StatusIndicator                          | `status`      |
+| :--------------------------------------- | :------------ |
+| <StatusIndicator status="success" />     | `success`     |
+| <StatusIndicator status="critical" />    | `critical`    |
+| <StatusIndicator status="informative" /> | `informative` |
+| <StatusIndicator status="warning" />     | `warning`     |
+| <StatusIndicator status="inactive" />    | `inactive`    |
