+++ packages/components/src/StatusIndicator/StatusIndicator.css.d.ts
@@ -1,0 +1,5 @@
+declare const styles: {
+  readonly "statusIndicator": string;
+};
+export = styles;
+
