+++ packages/components/src/StatusIndicator/StatusIndicator.tsx
@@ -1,0 +1,22 @@
+import React from "react";
+import styles from "./StatusIndicator.css";
+
+export type StatusType =
+  | "success"
+  | "warning"
+  | "critical"
+  | "inactive"
+  | "informative";
+
+interface StatusIndicatorProps {
+  readonly status: StatusType;
+}
+
+export function StatusIndicator({ status }: StatusIndicatorProps) {
+  return (
+    <div
+      style={{ backgroundColor: `var(--color-${status}` }}
+      className={styles.statusIndicator}
+    />
+  );
+}
