+++ packages/components/src/StatusIndicator/index.ts
@@ -1,0 +1,1 @@
+export { StatusIndicator } from "./StatusIndicator";
