+++ packages/components/src/StatusLabel/StatusLabel.css
@@ -10,15 +10,6 @@
   width: fit-content;
 }
 
-.statusIndicator {
-  width: var(--statusLabel-icon-diameter);
-  height: var(--statusLabel-icon-diameter);
-  margin-top: var(--space-minuscule);
-  border-radius: 3px;
-  background-color: var(--color-success);
-  flex-shrink: 0;
-}
-
 .labelTextEndAligned {
   flex-direction: row-reverse;
 }
