+++ packages/components/src/StatusLabel/StatusLabel.css.d.ts
@@ -1,6 +1,5 @@
 declare const styles: {
   readonly "statusLabelRow": string;
-  readonly "statusIndicator": string;
   readonly "labelTextEndAligned": string;
 };
 export = styles;
