+++ packages/components/src/StatusLabel/StatusType.ts
@@ -1,0 +1,6 @@
+export type StatusType =
+  | "success"
+  | "warning"
+  | "critical"
+  | "inactive"
+  | "informative";
