+++ packages/components/src/StatusLabel/index.ts
@@ -1,1 +1,2 @@
 export { StatusLabel } from "./StatusLabel";
+export type { StatusType } from "./StatusType";
